+++ versions.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     cato = {
       source  = "catonetworks/cato"
-      version = ">= 0.0.70"
+      version = "~> 0.0.93"
     }
     azurerm = {
       source  = "hashicorp/azurerm"